+++ src/main/java/com/esquad/esquadbe/qnaboard/service/CommentService.java
@@ -1,0 +1,4 @@
+package com.esquad.esquadbe.qnaboard.service;
+
+public class CommentService {
+}
