--- src/main/java/com/esquad/esquadbe/service/CommentService.java
@@ -1,4 +1,0 @@
-package com.esquad.esquadbe.service;
-
-public class CommentService {
-}
