--- src/main/java/com/esquad/esquadbe/service/QuestionService.java
@@ -1,9 +1,0 @@
-package com.esquad.esquadbe.service;
-
-
-import org.springframework.stereotype.Service;
-
-
-@Service
-public class QuestionService {
-}
