+++ src/main/resources/samples/user.sql
@@ -1,0 +1,14 @@
+INSERT INTO users (
+    user_id,
+    user_name,
+    nickname,
+    password,
+    email,
+    phone_no,
+    birth_day,
+    address
+) VALUES
+      ('user1', 'user1', 'user11', 'u3kvlk9ekd', 'user1@naver.com', '01091358452', '1992-05-04', '부산시 수영구 광안해변로 291'),
+      ('user2', 'user2', 'user12', 'fenlciok22', 'user2@naver.com', '01091351111', '1992-08-04', '부산시 수영구 광안해변로 291'),
+      ('user3', 'user3', 'user13', 'olenlcidhl2', 'user3@naver.com', '01022221111', '1996-08-04', '부산시 수영구 광안해변로 291'),
+      ('user4', 'user4', 'user14', 'i3bkcudb2ol', 'user4@naver.com', '01091315111', '1992-04-12', '부산시 수영구 광안해변로 291');
